Move DINOv2 work to dinov2_legacy; reserve dinov2 for canonical port
Rename paz.models.foundation.dinov2 to dinov2_legacy with git mv so the
existing standard and windowed DINOv2 work, detector utilities, and tests
stay intact and history is preserved. Rewrite internal imports to the new
package path. Create a fresh, empty paz.models.foundation.dinov2 package
reserved for the upcoming function-only canonical implementation.

No numerical behavior changes. Legacy tests: 42 passed, 83 skipped, 10
errors (pre-existing, missing precomputed output fixtures) before and
after the move.
