+++ paz/models/foundation/dinov2/__init__.py
@@ -1,11 +1,6 @@
-from paz.models.foundation.dinov2.models.vision_transformer import arange_indices
-from paz.models.foundation.dinov2.models.windowed_vision_transformer import (
-    WindowedDinov2PatchEmbeddings,
-    WindowedDinov2Layer,
-    WindowedDinov2Encoder,
-    WindowedDinov2Model,
-    dinov2_windowed_small,
-    dinov2_windowed_base,
-    dinov2_windowed_large,
-    dinov2_windowed_giant,
-)
+"""Canonical function-only DINOv2 implementation.
+
+Reserved for the new implementation built from reusable PAZ transformer
+primitives. The previous work lives in
+``paz.models.foundation.dinov2_legacy`` until the detector migrates.
+"""
