+++ paz/models/foundation/dinov2_legacy/ARCHITECTURE.md
