+++ paz/models/foundation/dinov2_legacy/__init__.py
@@ -1,0 +1,11 @@
+from paz.models.foundation.dinov2_legacy.models.vision_transformer import arange_indices
+from paz.models.foundation.dinov2_legacy.models.windowed_vision_transformer import (
+    WindowedDinov2PatchEmbeddings,
+    WindowedDinov2Layer,
+    WindowedDinov2Encoder,
+    WindowedDinov2Model,
+    dinov2_windowed_small,
+    dinov2_windowed_base,
+    dinov2_windowed_large,
+    dinov2_windowed_giant,
+)
