+++ paz/models/foundation/dinov2_legacy/layers/layer_scale.py
