+++ paz/models/foundation/dinov2_legacy/models/__init__.py
@@ -1,4 +1,4 @@
-from paz.models.foundation.dinov2.models.vision_transformer import (
+from paz.models.foundation.dinov2_legacy.models.vision_transformer import (
     DINOV2VIT,
     DINOV2Small,
     DINOV2Base,
