+++ paz/models/foundation/dinov2_legacy/test_dinov2_gaint_model_outputs.py
